requirements.txt
# Core Flask dependencies
Flask==1.1.2
Werkzeug==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
click==7.1.2

# Data processing and ML
numpy>=1.19.4,<1.24.0
pandas>=1.1.4,<1.6.0
scikit-learn>=0.23.2,<0.25.0
scipy>=1.5.4,<1.11.0
joblib>=0.17.0
threadpoolctl>=2.1.0

# PyTorch - install separately if this fails (see instructions below)
torch>=1.7.0
torchvision>=0.8.1
torchaudio>=0.7.0

# Image processing
Pillow>=8.0.1,<10.0.0

# HTTP requests
requests>=2.25.0
urllib3>=1.26.2,<2.0.0
certifi>=2020.12.5
idna>=2.10
chardet>=3.0.4

# Server
gunicorn==20.0.4

# Environment variables (.env)
python-dotenv>=0.19.0

# Utilities
python-dateutil>=2.8.1
pytz>=2020.4
six>=1.15.0
packaging>=20.4
attrs>=20.3.0

# Optional: Jupyter and development tools (uncomment if needed)
# ipykernel>=5.3.4
# ipython>=7.19.0
# ipython-genutils>=0.2.0
# ipywidgets>=7.5.1
# jupyter>=1.0.0
# jupyter-client>=6.1.7
# jupyter-console>=6.2.0
# jupyter-core>=4.7.0
# notebook>=6.1.5
# pyzmq>=20.0.0,<25.0.0
# tornado>=6.1
# traitlets>=5.0.5
# nbclient>=0.5.1
# nbconvert>=6.0.7
# nbformat>=5.0.8
# nest-asyncio>=1.4.3
# prometheus-client>=0.9.0
# terminado>=0.9.1

# Optional: Visualization (uncomment if needed)
# matplotlib>=3.3.3,<3.8.0
# seaborn>=0.11.0
# cycler>=0.10.0
# kiwisolver>=1.3.1
# pyparsing>=2.4.7

# Optional: Code quality tools (uncomment if needed)
# pylint>=2.6.0
# astroid>=2.4.2
# isort>=5.6.4
# pycodestyle>=2.6.0
# pyflakes>=2.2.0
# mccabe>=0.6.1
# lazy-object-proxy>=1.4.3
# wrapt>=1.12.1

# Other utilities
# argon2-cffi>=20.1.0
# cffi>=1.14.4
# pycparser>=2.20
# simplejson>=3.17.2
# typing-extensions>=3.7.4.3

# IMPORTANT: If PyTorch installation fails, install it separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
